Feat:The responses of the Client's enpoints are corrected

+++ src/clientes/clientes.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Cliente } from '../entities/cliente.entity';
@@ -15,7 +19,13 @@ export class ClientesService {
   }
 
   async findByName(nombre: string): Promise<Cliente[]> {
-    return this.clienteRepository.find({ where: { business_name: nombre } });
+    const cliente = await this.clienteRepository.find({
+      where: { business_name: nombre },
+    });
+    if (cliente.length === 0) {
+      throw new NotFoundException('Cliente no encontrado');
+    }
+    return cliente;
   }
 
   async findRFCAndBusinessName(): Promise<Partial<Cliente[]>> {
@@ -25,18 +35,39 @@ export class ClientesService {
   }
 
   async findOne(rfc: string): Promise<Cliente> {
-    return this.clienteRepository.findOne({ where: { rfc } });
+    const cliente = await this.clienteRepository.findOne({ where: { rfc } });
+    if (!cliente) {
+      throw new NotFoundException('Cliente no encontrado');
+    }
+    return cliente;
   }
 
   async create(cliente: Cliente): Promise<Cliente> {
+    const rfc = cliente.rfc;
+    const existingCliente = await this.clienteRepository.findOne({
+      where: { rfc },
+    });
+    console.log(existingCliente);
+    if (existingCliente) {
+      throw new ConflictException('Cliente ya existente');
+    }
+
     return this.clienteRepository.save(cliente);
   }
 
   async update(rfc: string, cliente: Cliente): Promise<void> {
+    const existingCliente = await this.findOne(rfc);
+    if (!existingCliente) {
+      throw new ConflictException('Cliente no existente');
+    }
     await this.clienteRepository.update({ rfc }, cliente);
   }
 
   async remove(rfc: string): Promise<void> {
+    const existingCliente = await this.findOne(rfc);
+    if (!existingCliente) {
+      throw new ConflictException('Cliente no existente');
+    }
     await this.clienteRepository.delete({ rfc });
   }
 }
